CodeGenのぶり対応拡張セットを使ったサンプルを組み込んでみた ぶり対応拡張セットをテストするために作ったやつをそのままとか・・・
git-svn-id: https://www.seasar.org/svn/sandbox/s2buri/trunk@144 1c266114-290a-0410-b1ce-a63319e86887

+++ Buri2.3/src/test/java/example/org/seasar/buri/test/PrivacyInfoManagementTest.java
@@ -65,7 +65,7 @@ public class PrivacyInfoManagementTest extends S2TestCase {
     }
     
     private void setupBuriTestUser(long userID) {
-        BuriTestUserDto testUserDto = userDao.getBuriTestUser(userID);
+        BuriTestUserDto testUserDto = userDao.getBuriTestUser((int)userID);
         userDto.setUserID(testUserDto.getUserID());
         userDto.setUserName(testUserDto.getUserName());
     }
